% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{append_slope_features}
\alias{append_slope_features}
\title{Slope and Elevation Features}
\usage{
append_slope_features(DT, hostName = "localhost", port = 5005)
}
\arguments{
\item{DT}{A \code{\link{data.table}} with at least the following columns:
\itemize{
    \item{\emph{latitude}}{: latitude in degrees}
    \item{\emph{longitude}}{: longitude in degrees}
    \item{\emph{assetId}}{: identifier for your asset. Requests are separated
                          by assetId to avoid computing features like \code{bearing}
                          over data from different physical things}
    \item{\emph{bearing}}{: (optional). If given, this should be the final bearing
                         in degrees. If not supplied, this will be calculated
                         from consecutive observations}
}}

\item{hostName}{A string with the host running groundhog. By default, this
function expects that you're running the app on \code{localhost}.}

\item{port}{Port that the service is running on. 5005 by default. You PROBABLY
won't ever have to change this.}
}
\value{
Nothing. This function will modify \code{DT} in place by appending columns
}
\description{
This function takes a \code{\link{data.table}} with GPS coordinates
             and enriches it with some elevation and slope features. The function
             hits the public Shuttle Radar Topography Mission (SRTM) dataset
             to get elevation data for each lat-lon pair in your dataset.
}
\examples{
\dontrun{
library(data.table)

# Create a sample dataset
someDT <- data.table::data.table(
     longitude = runif(10, -110, -109)
     , latitude = runif(10, 45, 46)
     , dateTime = seq.POSIXt(from = as.POSIXct("2017-01-01", tz = "UTC")
                             , to = as.POSIXct("2017-01-15", tz = "UTC")
                             , length.out = 10)
     , assetId = c(rep("ABC", 5), rep("DEF", 5))
)

# Append slope fearures
groundhog::append_slope_features(someDT, hostName = "localhost", port = 5005)
}
}
\references{
\href{https://www2.jpl.nasa.gov/srtm/}{Background on SRTM:}
}
