% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{concentration}
\alias{concentration}
\title{Define a concentration object to be passed to the Rx Studio API}
\usage{
concentration(datetime, concentration)
}
\arguments{
\item{datetime}{date and time of the administered dose}

\item{concentration}{Concentration (number)}
}
\value{
list
}
\description{
Define a concentration object to be passed to the Rx Studio API
}
\examples{
concentration(Sys.time(), 10)
}
