% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dose}
\alias{dose}
\title{Define a dose object to be passed to the Rx Studio API}
\usage{
dose(datetime, dose, tinf = NULL, route = c("IV", "oral"))
}
\arguments{
\item{datetime}{date and time of the administered dose}

\item{dose}{Dose Amount (number in mg)}

\item{tinf}{optional Infusion Time (number in hours)}

\item{route}{Route of Administration, e.g. \code{IV} or \code{oral}}
}
\value{
list
}
\description{
Define a dose object to be passed to the Rx Studio API
}
\examples{
dose(Sys.time(), 2000, 2)
dose(Sys.time(), 2000, route = 'oral')
}
