% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_id_token}
\alias{get_id_token}
\title{Get an id token using the access token}
\usage{
get_id_token()
}
\arguments{
\item{access_token}{string}
}
\description{
Get an id token using the access token
}
\keyword{internal}
