% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{print.rx_studio_report}
\alias{print.rx_studio_report}
\title{Open HTML in browser}
\usage{
\method{print}{rx_studio_report}(x, ...)
}
\arguments{
\item{x}{HTML document returned by \code{simulate}}

\item{...}{further parameters}
}
\description{
Open HTML in browser
}
