% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{read_cached_simulation}
\alias{read_cached_simulation}
\title{Load previously run Rx Studio simulation}
\usage{
read_cached_simulation(request_id, format = c("json", "html"))
}
\arguments{
\item{request_id}{UUID}
}
\value{
\code{list} or HTML
}
\description{
Load previously run Rx Studio simulation
}
