% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{regimen}
\alias{regimen}
\title{Define a dosing regimen object to be passed to the Rx Studio API}
\usage{
regimen(dose, interval = NULL, tinf = NULL, type = c("REGIMEN", "ORALREGIMEN"))
}
\arguments{
\item{dose}{Dose Amount (number in mg)}

\item{interval}{optional Dosing Interval (number in hours)}

\item{tinf}{optional Infusion Time (number in hours)}

\item{type}{Regimen type, e.g. \code{REGIMEN} or \code{ORALREGIMEN}}
}
\value{
list
}
\description{
Define a dosing regimen object to be passed to the Rx Studio API
}
\examples{
regimen(2000)
regimen(2000, 8, 2)
regimen(2000, 12, type = 'ORALREGIMEN')
}
