% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_access_token}
\alias{set_access_token}
\title{Sets access token for Rx Studio API usage}
\usage{
set_access_token(access_token)
}
\arguments{
\item{access_token}{token generated at \url{https://app.rx.studio/profile/preferences}}
}
\description{
Sets access token for Rx Studio API usage
}
