% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{simulate}
\alias{simulate}
\title{Hit Rx Studio simulation API}
\usage{
simulate(endpoint, ...)
}
\arguments{
\item{endpoint}{path under \url{https://api.rx.studio/v1/simulation/}}

\item{...}{named parameters passed}
}
\description{
Hit Rx Studio simulation API
}
\examples{
\dontrun{
simulate('cefepime_mc_pta_ftime_above_mic',
  PATID = 'Anonymous', AGE = 37, HEIGHT = 175, WEIGHT = 90, GENDER = 'MALE',
  CREATININE = 1, MODEL = 'Tam et al. (2003) - General ward',
  PCTABOVEMIC = 60, CRCLCAP = 'No cap', REGIMENS = list(regimen(1000, 12, 2)))
}
}
\keyword{internal}
