% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_amikacin_mc_cfr_cmax_mic_ratio.R
\name{simulate_amikacin_mc_cfr_cmax_mic_ratio}
\alias{simulate_amikacin_mc_cfr_cmax_mic_ratio}
\title{Amikacin » CFR » Cmax/MIC}
\usage{
simulate_amikacin_mc_cfr_cmax_mic_ratio(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  MODEL,
  CREATININE,
  CMAXPERMIC,
  REGIMENS,
  MICDISTR
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Saez Fernandez et al. (2019) - General ward').}

\item{CREATININE}{Creatinine. Serum creatinine. Must be provided as numeric (min. 0.01, max. 15 mg/dL).}

\item{CMAXPERMIC}{Maximum concentration to MIC ratio target. The PK/PD target can be provided as maximum concentration to minimum inhibitory concentration ratio (Cmax/MIC). Must be provided as numeric (min. 0.1, max. 200 ).}

\item{REGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-8 'REGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}

\item{MICDISTR}{MIC distribution. Non-cumulative MIC distribution of organism for the given antibiotic at the specified MICs provided as percentages with a sum of 100\%. MIC distribution of amikacin and comparator antimicrobial agents when tested against Pseudomonas isolates in the SENTRY program collected during 2013-2019 (mg/L). Dataset provided by JMI Laboratories and the SENTRY Antimicrobial Surveillance Program, available at <a href="http://sentry-mvp.jmilabs.com" target="_new">sentry-mvp.jmilabs.com</a>. Must be provided as set.}
}
\description{
Amikacin » Cumulative Fraction of Response » Total maximum concentration to MIC ratio
}
\details{
\strong{Drug}:
Amikacin

\strong{Method}:
Monte Carlo simulation on the proportion of the population achieving a certain pharmacodynamic index value, given the minimum inhibitory concentration (MIC) distribution of the target microorganism(s).

\strong{PK/PD target}:
Maximum concentration (mg/L) to minimum inhibitory concentration ratio (mg/L).
}
\examples{
\dontrun{
simulate_amikacin_mc_cfr_cmax_mic_ratio(PATID = "Anonymous", 
    AGE = 75, HEIGHT = 160, 
    WEIGHT = 105, GENDER = "Male", 
    MODEL = "Saez Fernandez et al. (2019) - General ward", 
    CREATININE = 1, CMAXPERMIC = 8, 
    REGIMENS = list(list(
        DOSE = 400, INTERVAL = 4, 
        TINF = 0.5, set = "REGIMEN")), 
    MICDISTR = list("MICDISTR-0.25" = list(
        0.8), "MICDISTR-0.5" = list(
        2.7), "MICDISTR-1" = list(
        6.4), "MICDISTR-2" = list(
        29.8), "MICDISTR-4" = list(
        34.3), "MICDISTR-8" = list(
        13.6), "MICDISTR-16" = list(
        4.8), "MICDISTR-32" = list(
        2.3), "MICDISTR-64" = list(
        5.3)))
}

}
\references{
\itemize{
  \item Saez Fernandez et al. (2019): Evaluation of renal function equations to predict amikacin clearance. In. Expert Review of Clinical Pharmacology. https://www.tandfonline.com/doi/full/10.1080/17512433.2019.1637253
}
}
