% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_amikacin_mc_pta_cmin_mic_ratio.R
\name{simulate_amikacin_mc_pta_cmin_mic_ratio}
\alias{simulate_amikacin_mc_pta_cmin_mic_ratio}
\title{Amikacin » PTA » Cmin/MIC}
\usage{
simulate_amikacin_mc_pta_cmin_mic_ratio(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  MODEL,
  CREATININE,
  CMINPERMIC,
  REGIMENS
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Sex of the patient. Must be provided as string ('Male' or 'Female').}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Saez Fernandez et al. (2019) - General ward').}

\item{CREATININE}{Creatinine. Serum creatinine. Must be provided as numeric (min. 0.01, max. 15 mg/dL).}

\item{CMINPERMIC}{Minimum concentration to MIC ratio target. The PK/PD target can be provided as minimum concentration to minimum inhibitory concentration ratio (Cmin/MIC). Must be provided as numeric (min. 0.1, max. 50 ).}

\item{REGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-8 'REGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}
}
\description{
Amikacin » Probability of Target Attainment » Total minimum concentration to MIC ratio
}
\details{
\strong{Drug}:
Amikacin

\strong{Method}:
Monte Carlo simulation on the probabilities that a specific value of a pharmacodynamic index is achieved in case of different minimum inhibitory concentrations (MIC).

\strong{PK/PD target}:
Minimum concentration (mg/L) to minimum inhibitory concentration ratio (mg/L).
}
\examples{
\dontrun{
simulate_amikacin_mc_pta_cmin_mic_ratio(PATID = "Anonymous", 
    AGE = 35, HEIGHT = 155, 
    WEIGHT = 55, GENDER = "Female", 
    MODEL = "Saez Fernandez et al. (2019) - General ward", 
    CREATININE = 1.2, 
    CMINPERMIC = 6, REGIMENS = list(
        list(DOSE = 200, 
            INTERVAL = 6, 
            TINF = 0.5, 
            set = "REGIMEN")))
}

}
\references{
\itemize{
  \item Saez Fernandez et al. (2019): Evaluation of renal function equations to predict amikacin clearance. In. Expert Review of Clinical Pharmacology. https://www.tandfonline.com/doi/full/10.1080/17512433.2019.1637253
}
}
