% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cefepime_mc_cfr_ftime_above_mic.R
\name{simulate_cefepime_mc_cfr_ftime_above_mic}
\alias{simulate_cefepime_mc_cfr_ftime_above_mic}
\title{Cefepime » CFR » % fT > MIC}
\usage{
simulate_cefepime_mc_cfr_ftime_above_mic(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  CREATININE,
  MODEL,
  PCTABOVEMIC,
  CRCLCAP,
  REGIMENS,
  MICDISTR
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Sex of the patient. Must be provided as string ('Male' or 'Female').}

\item{CREATININE}{Creatinine. Serum creatinine. Must be provided as numeric (min. 0.01, max. 15 mg/dL).}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Nicasio et al. (2009) - ICU' or 'Tam et al. (2003) - General ward').}

\item{PCTABOVEMIC}{Percent of time target that the drug concentration is above MIC. The PK/PD target can be provided as the percent of time that the drug concentration is above the minimum inhibitory concentration (% T > MIC). Must be provided as numeric (min. 5, max. 100 %).}

\item{CRCLCAP}{Capping Creatinine Clearance. Whether to use capping for creatinine clearance. Must be provided as string ('No cap', '120 ml/min', '130 ml/min', '140 ml/min' or '150 ml/min').}

\item{REGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-8 'REGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}

\item{MICDISTR}{MIC distribution. Non-cumulative MIC distribution of organism for the given antibiotic at the specified MICs provided as percentages with a sum of 100%. MIC distribution of cefepime and comparator antimicrobial agents when tested against Enterobacter isolates in the SENTRY program collected during 2018 (mg/L). Dataset provided by JMI Laboratories and the SENTRY Antimicrobial Surveillance Program, available at <a href="http://sentry-mvp.jmilabs.com" target="_new">sentry-mvp.jmilabs.com</a>. Must be provided as set.}
}
\description{
Cefepime » Cumulative Fraction of Response » Percent time of free concentration above MIC
}
\section{Drug}{

Cefepime
}

\section{Method}{

Monte Carlo simulation on the proportion of the population achieving a certain pharmacodynamic index value, given the minimum inhibitory concentration (MIC) distribution of the target microorganism(s).
}

\section{PK/PD target}{

The percent of time that the free concentration is above the minimum inhibitory concentration.
}

\examples{
\dontrun{
simulate_cefepime_mc_cfr_ftime_above_mic(PATID = "Anonymous", 
    AGE = 75, HEIGHT = 160, 
    WEIGHT = 105, GENDER = "Male", 
    CREATININE = 1, MODEL = "Nicasio et al. (2009) - ICU", 
    PCTABOVEMIC = 60, 
    CRCLCAP = "No cap", 
    REGIMENS = list(list(
        DOSE = 1000, 
        INTERVAL = 8, 
        TINF = 0.5, set = "REGIMEN")), 
    MICDISTR = list("MICDISTR-0.0078125" = list(
        0), "MICDISTR-0.015625" = list(
        0.8), "MICDISTR-0.03125" = list(
        21.4), "MICDISTR-0.0625" = list(
        33.9), "MICDISTR-0.125" = list(
        12.8), "MICDISTR-0.25" = list(
        6), "MICDISTR-0.5" = list(
        3.5), "MICDISTR-1" = list(
        3.8), "MICDISTR-2" = list(
        5), "MICDISTR-4" = list(
        3.5), "MICDISTR-8" = list(
        2), "MICDISTR-16" = list(
        1.2), "MICDISTR-32" = list(
        1), "MICDISTR-64" = list(
        1.7), "MICDISTR-128" = list(
        0.8), "MICDISTR-256" = list(
        0.9), "MICDISTR-512" = list(
        1.7)))
}
}
