% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ceftriaxone_bayesian_fcmin_mic_ratio.R
\name{simulate_ceftriaxone_bayesian_fcmin_mic_ratio}
\alias{simulate_ceftriaxone_bayesian_fcmin_mic_ratio}
\title{Ceftriaxone » Bayesian » fCmin/MIC}
\usage{
simulate_ceftriaxone_bayesian_fcmin_mic_ratio(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  MODEL,
  MIC,
  CMINPERMIC,
  CRCLCAP,
  HISTORY,
  REGIMENS
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Sex of the patient. Must be provided as string ('Male' or 'Female').}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Garot et al. (2011) - ICU, septic patients').}

\item{MIC}{MIC. Minimum Inhibitory Concentration (MIC). Must be provided as numeric (min. 0.01, max. 1024 mg/L).}

\item{CMINPERMIC}{Minimum concentration to MIC ratio target. The PK/PD target can be provided as minimum concentration to minimum inhibitory concentration ratio (Cmin/MIC). Must be provided as numeric (min. 0.1, max. 50 ).}

\item{CRCLCAP}{Capping Creatinine Clearance. Whether to use capping for creatinine clearance. Must be provided as string ('No cap', '120 ml/min', '130 ml/min', '140 ml/min' or '150 ml/min').}

\item{HISTORY}{Historical Records.  Must be provided as list of 3-24 'HISTCREATININE', 'HISTDOSE' or 'HISTCONCENTRATION' values.}

\item{REGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-20 'REGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}
}
\description{
Ceftriaxone » Bayesian adaptive dosing » Free minimum concentration to MIC ratio
}
\section{Drug}{

Ceftriaxone
}

\section{Method}{

Estimate the pharmacokinetic parameters of the patient from past concentrations and creatinine levels with Bayesian inverse modeling, then use that information to predict the steady state concentrations for multiple dosing regimens and select the optimal one, with regard to the target pharmacodynamic index.
}

\section{PK/PD target}{

Free minimum concentration (mg/L) to minimum inhibitory concentration ratio (mg/L).
}

\examples{
\dontrun{
simulate_ceftriaxone_bayesian_fcmin_mic_ratio(PATID = "Anonymous", 
    AGE = 65, HEIGHT = 175, 
    WEIGHT = 75, GENDER = "Male", 
    MODEL = "Garot et al. (2011) - ICU, septic patients", 
    MIC = 2, CMINPERMIC = 1, 
    CRCLCAP = "No cap", 
    HISTORY = list(list(
        DATETIME = structure(1601870400, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
        DOSE = 2000, 
        TINF = 0.5, set = "HISTDOSE"), 
        list(DATETIME = structure(1601866800, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            CREATININE = 0.9, 
            set = "HISTCREATININE"), 
        list(DATETIME = structure(1601881200, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            CONCENTRATION = 14, 
            set = "HISTCONCENTRATION"), 
        list(DATETIME = structure(1601892000, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            CONCENTRATION = 10, 
            set = "HISTCONCENTRATION")), 
    REGIMENS = list(list(
        DOSE = 500, INTERVAL = 12, 
        TINF = 0.5, set = "REGIMEN"), 
        list(DOSE = 500, 
            INTERVAL = 24, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 1000, 
            INTERVAL = 12, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 1000, 
            INTERVAL = 24, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 1500, 
            INTERVAL = 12, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 1500, 
            INTERVAL = 24, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 2000, 
            INTERVAL = 12, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 2000, 
            INTERVAL = 24, 
            TINF = 0.5, 
            set = "REGIMEN")))
}
}
