% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ceftriaxone_mc_pta_ftime_above_mic.R
\name{simulate_ceftriaxone_mc_pta_ftime_above_mic}
\alias{simulate_ceftriaxone_mc_pta_ftime_above_mic}
\title{Ceftriaxone » PTA » \% fT > MIC}
\usage{
simulate_ceftriaxone_mc_pta_ftime_above_mic(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  CREATININE,
  MODEL,
  PCTABOVEMIC,
  CRCLCAP,
  REGIMENS
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').}

\item{CREATININE}{Creatinine. Serum creatinine. Must be provided as numeric (min. 0.01, max. 15 mg/dL).}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Garot et al. (2011) - ICU, septic patients').}

\item{PCTABOVEMIC}{Percent of time target that the drug concentration is above MIC. The PK/PD target can be provided as the percent of time that the drug concentration is above the minimum inhibitory concentration (\% T > MIC). Must be provided as numeric (min. 5, max. 100 \%).}

\item{CRCLCAP}{Capping Creatinine Clearance. Whether to use capping for creatinine clearance. Must be provided as string ('No cap', '120 ml/min', '130 ml/min', '140 ml/min' or '150 ml/min').}

\item{REGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-8 'REGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}
}
\description{
Ceftriaxone » Probability of Target Attainment » Percent time of free concentration above MIC
}
\details{
\strong{Drug}:
Ceftriaxone

\strong{Method}:
Monte Carlo simulation on the probabilities that a specific value of a pharmacodynamic index is achieved in case of different minimum inhibitory concentrations (MIC).

\strong{PK/PD target}:
The percent of time that the free concentration is above the minimum inhibitory concentration.
}
\examples{
\dontrun{
simulate_ceftriaxone_mc_pta_ftime_above_mic(PATID = "Anonymous", 
    AGE = 55, HEIGHT = 195, 
    WEIGHT = 95, GENDER = "Male", 
    CREATININE = 1, MODEL = "Garot et al. (2011) - ICU, septic patients", 
    PCTABOVEMIC = 90, 
    CRCLCAP = "No cap", 
    REGIMENS = list(list(
        DOSE = 500, INTERVAL = 12, 
        TINF = 0.5, set = "REGIMEN"), 
        list(DOSE = 500, 
            INTERVAL = 24, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 1000, 
            INTERVAL = 12, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 1000, 
            INTERVAL = 24, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 1500, 
            INTERVAL = 12, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 1500, 
            INTERVAL = 24, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 2000, 
            INTERVAL = 12, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 2000, 
            INTERVAL = 24, 
            TINF = 0.5, 
            set = "REGIMEN")))
}

}
\references{
\itemize{
  \item Garot, Denis et al. (2011): Population pharmacokinetics of ceftriaxone in critically ill septic patients: a reappraisal. In. British Journal of Clinical Pharmacology. https://bpspubs.onlinelibrary.wiley.com/doi/full/10.1111/j.1365-2125.2011.04005.x
}
}
