% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gentamicin_mc_cfr_auc_mic_ratio.R
\name{simulate_gentamicin_mc_cfr_auc_mic_ratio}
\alias{simulate_gentamicin_mc_cfr_auc_mic_ratio}
\title{Gentamicin » CFR » AUC/MIC}
\usage{
simulate_gentamicin_mc_cfr_auc_mic_ratio(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  MODEL,
  CREATININE,
  AUCPERMIC,
  REGIMENS,
  MICDISTR
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Pai et al. (2011) - General ward').}

\item{CREATININE}{Creatinine. Serum creatinine. Must be provided as numeric (min. 0.01, max. 15 mg/dL).}

\item{AUCPERMIC}{AUC to MIC ratio target. The PK/PD target can be provided as 24 hour area under the concentration-time curve to minimum inhibitory concentration ratio (AUC/MIC). Must be provided as numeric (min. 10, max. 2000 ).}

\item{REGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-8 'REGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}

\item{MICDISTR}{MIC distribution. Non-cumulative MIC distribution of organism for the given antibiotic at the specified MICs provided as percentages with a sum of 100\%. MIC distribution of gentamicin and comparator antimicrobial agents when tested against Pseudomonas isolates in the SENTRY program collected during 2019 (mg/L). Dataset provided by JMI Laboratories and the SENTRY Antimicrobial Surveillance Program, available at <a href="http://sentry-mvp.jmilabs.com" target="_new">sentry-mvp.jmilabs.com</a>. Must be provided as set.}
}
\description{
Gentamicin » Cumulative Fraction of Response » AUC to MIC ratio
}
\details{
\strong{Drug}:
Gentamicin

\strong{Method}:
Monte Carlo simulation on the proportion of the population achieving a certain pharmacodynamic index value, given the minimum inhibitory concentration (MIC) distribution of the target microorganism(s).

\strong{PK/PD target}:
24 hour area under the concentration-time curve to minimum inhibitory concentration ratio.
}
\examples{
\dontrun{
simulate_gentamicin_mc_cfr_auc_mic_ratio(PATID = "Anonymous", 
    AGE = 75, HEIGHT = 160, 
    WEIGHT = 105, GENDER = "Male", 
    MODEL = "Pai et al. (2011) - General ward", 
    CREATININE = 0.8760563, 
    AUCPERMIC = 50, REGIMENS = list(
        list(DOSE = 300, 
            INTERVAL = 6, 
            TINF = 0.5, 
            set = "REGIMEN")), 
    MICDISTR = list("MICDISTR-0.125" = list(
        2.2), "MICDISTR-0.25" = list(
        1.5), "MICDISTR-0.5" = list(
        4), "MICDISTR-1" = list(
        25.4), "MICDISTR-2" = list(
        38.7), "MICDISTR-4" = list(
        13.7), "MICDISTR-8" = list(
        4.8), "MICDISTR-16" = list(
        2), "MICDISTR-32" = list(
        7.6)))
}

}
\references{
\itemize{
  \item Pai et al. (2011): Simplified Estimation of Aminoglycoside Pharmacokinetics in Underweight and Obese Adult Patients. In. Antimicrobial Agents and Chemotherapy. https://aac.asm.org/content/55/9/4006
}
}
