% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gentamicin_mc_emp_cmin_cmax.R
\name{simulate_gentamicin_mc_emp_cmin_cmax}
\alias{simulate_gentamicin_mc_emp_cmin_cmax}
\title{Gentamicin » Empiric » Cmin & Cmax}
\usage{
simulate_gentamicin_mc_emp_cmin_cmax(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  MODEL,
  CREATININE,
  MIC,
  CMIN,
  CMAX,
  LOADINGDOSE,
  REGIMENS
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Pai et al. (2011) - General ward').}

\item{CREATININE}{Creatinine. Serum creatinine. Must be provided as numeric (min. 0.01, max. 15 mg/dL).}

\item{MIC}{MIC. Minimum Inhibitory Concentration (MIC). Must be provided as numeric (min. 0.01, max. 1024 mg/L).}

\item{CMIN}{Minimum concentration target. The PK/PD target can be provided as minimum blood plasma concentration (Cmin). For selecting an appropriate target please refer to the dosing guideline of your institution, e.g. <a href="http://med.stanford.edu/bugsanddrugs/guidebook/_jcr_content/main/panel_builder_584648957/panel_0/download_1194988017/file.res/Aminoglycoside\%20Dosing\%20Guide\%202019-05-20.pdf" target="_new">Stanford Health Care Aminoglycoside Dosing Guide</a>. Must be provided as numeric (min. 0.1, max. 200 mg/L).}

\item{CMAX}{Maximum concentration target. The PK/PD target can be provided as maximum blood plasma concentration (Cmax). For selecting an appropriate target please refer to the dosing guideline of your institution, e.g. <a href="http://med.stanford.edu/bugsanddrugs/guidebook/_jcr_content/main/panel_builder_584648957/panel_0/download_1194988017/file.res/Aminoglycoside\%20Dosing\%20Guide\%202019-05-20.pdf" target="_new">Stanford Health Care Aminoglycoside Dosing Guide</a>. Must be provided as numeric (min. 0.1, max. 200 mg/L).}

\item{LOADINGDOSE}{Loading dose. Loading dose is desired or not. Must be provided as string ('No' or 'Yes').}

\item{REGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-20 'REGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}
}
\description{
Gentamicin » Empiric therapy » Total minimum and maximum concentration
}
\details{
\strong{Drug}:
Gentamicin

\strong{Method}:
Simulate concentrations for multiple dosing regimens and select the optimal one, with regard to the target pharmacodynamic index.

\strong{PK/PD target}:
Minimum and maximum blood plasma concentration (mg/L).
}
\examples{
\dontrun{
simulate_gentamicin_mc_emp_cmin_cmax(PATID = "Anonymous", 
    AGE = 65, HEIGHT = 175, 
    WEIGHT = 75, GENDER = "Male", 
    MODEL = "Pai et al. (2011) - General ward", 
    CREATININE = 0.876056338, 
    MIC = 1, CMIN = 2, 
    CMAX = 10, LOADINGDOSE = "No", 
    REGIMENS = list(list(
        set = "REGIMEN", 
        DOSE = 100, INTERVAL = 8, 
        TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 100, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 125, 
            INTERVAL = 8, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 125, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 150, 
            INTERVAL = 8, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 150, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 175, 
            INTERVAL = 8, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 175, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 200, 
            INTERVAL = 8, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 200, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 400, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 400, 
            INTERVAL = 36, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 500, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 500, 
            INTERVAL = 36, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 600, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 600, 
            INTERVAL = 36, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 700, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 700, 
            INTERVAL = 36, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 800, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 800, 
            INTERVAL = 36, 
            TINF = 0.5)))
}

}
\references{
\itemize{
  \item Pai et al. (2011): Simplified Estimation of Aminoglycoside Pharmacokinetics in Underweight and Obese Adult Patients. In. Antimicrobial Agents and Chemotherapy. https://aac.asm.org/content/55/9/4006
}
}
