% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_meropenem_mc_emp_fauc_mic_ratio.R
\name{simulate_meropenem_mc_emp_fauc_mic_ratio}
\alias{simulate_meropenem_mc_emp_fauc_mic_ratio}
\title{Meropenem » Empiric » fAUC/MIC}
\usage{
simulate_meropenem_mc_emp_fauc_mic_ratio(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  CREATININE,
  MODEL,
  EDEMA,
  MIC,
  AUCPERMIC,
  LOADINGDOSE,
  CRCLCAP,
  REGIMENS
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Sex of the patient. Must be provided as string ('Male' or 'Female').}

\item{CREATININE}{Creatinine. Serum creatinine. Must be provided as numeric (min. 0.01, max. 15 mg/dL).}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Crandon et al. (2011) - ICU', 'Li, C. et. al. (2006) - General ward' or 'Doh, K. et al. (2010) - Burn patients').}

\item{EDEMA}{Edema. Presence of edema in case of Burn Patients. Must be provided as string ('No' or 'Yes').}

\item{MIC}{MIC. Minimum Inhibitory Concentration (MIC). Must be provided as numeric (min. 0.01, max. 1024 mg/L).}

\item{AUCPERMIC}{AUC to MIC ratio target. The PK/PD target can be provided as 24 hour area under the concentration-time curve to minimum inhibitory concentration ratio (AUC/MIC). Must be provided as numeric (min. 10, max. 2000 ).}

\item{LOADINGDOSE}{Loading dose. Loading dose is desired or not. Must be provided as string ('No' or 'Yes').}

\item{CRCLCAP}{Capping Creatinine Clearance. Whether to use capping for creatinine clearance. Must be provided as string ('No cap', '120 ml/min', '130 ml/min', '140 ml/min' or '150 ml/min').}

\item{REGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-20 'REGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}
}
\description{
Meropenem » Empiric therapy » fAUC to MIC ratio
}
\section{Drug}{

Meropenem
}

\section{Method}{

Simulate concentrations for multiple dosing regimens and select the optimal one, with regard to the target pharmacodynamic index.
}

\section{PK/PD target}{

24 hour area under the free concentration-time curve to minimum inhibitory concentration ratio.
}

\examples{
\dontrun{
simulate_meropenem_mc_emp_fauc_mic_ratio(PATID = "Anonymous", 
    AGE = 65, HEIGHT = 175, 
    WEIGHT = 75, GENDER = "Male", 
    CREATININE = 0.876056338, 
    MODEL = "Crandon et al. (2011) - ICU", 
    EDEMA = "No", MIC = 1, 
    AUCPERMIC = 300, 
    LOADINGDOSE = "No", 
    CRCLCAP = "No cap", 
    REGIMENS = list(list(
        set = "REGIMEN", 
        DOSE = 500, INTERVAL = 6, 
        TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 500, 
            INTERVAL = 8, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 500, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 500, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 1000, 
            INTERVAL = 6, 
            TINF = 1), 
        list(set = "REGIMEN", 
            DOSE = 1000, 
            INTERVAL = 8, 
            TINF = 1), 
        list(set = "REGIMEN", 
            DOSE = 1000, 
            INTERVAL = 12, 
            TINF = 1), 
        list(set = "REGIMEN", 
            DOSE = 1000, 
            INTERVAL = 24, 
            TINF = 1), 
        list(set = "REGIMEN", 
            DOSE = 2000, 
            INTERVAL = 8, 
            TINF = 2), 
        list(set = "REGIMEN", 
            DOSE = 2000, 
            INTERVAL = 12, 
            TINF = 2), 
        list(set = "REGIMEN", 
            DOSE = 2000, 
            INTERVAL = 24, 
            TINF = 2)))
}
}
