% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_meropenem_mc_pta_ftime_above_mic.R
\name{simulate_meropenem_mc_pta_ftime_above_mic}
\alias{simulate_meropenem_mc_pta_ftime_above_mic}
\title{Meropenem » PTA » % fT > MIC}
\usage{
simulate_meropenem_mc_pta_ftime_above_mic(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  CREATININE,
  MODEL,
  EDEMA,
  PCTABOVEMIC,
  CRCLCAP,
  REGIMENS
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').}

\item{CREATININE}{Creatinine. Serum creatinine. Must be provided as numeric (min. 0.01, max. 15 mg/dL).}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Crandon et al. (2011) - ICU', 'Li, C. et. al. (2006) - General ward' or 'Doh, K. et al. (2010) - Burn patients').}

\item{EDEMA}{Edema. Presence of edema in case of Burn Patients. Must be provided as string ('No' or 'Yes').}

\item{PCTABOVEMIC}{Percent of time target that the drug concentration is above MIC. The PK/PD target can be provided as the percent of time that the drug concentration is above the minimum inhibitory concentration (% T > MIC). Must be provided as numeric (min. 5, max. 100 %).}

\item{CRCLCAP}{Capping Creatinine Clearance. Whether to use capping for creatinine clearance. Must be provided as string ('No cap', '120 ml/min', '130 ml/min', '140 ml/min' or '150 ml/min').}

\item{REGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-8 'REGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}
}
\description{
Meropenem » Probability of Target Attainment » Percent time of free concentration above MIC
}
\details{
\strong{Drug}:
Meropenem

\strong{Method}:
Monte Carlo simulation on the probabilities that a specific value of a pharmacodynamic index is achieved in case of different minimum inhibitory concentrations (MIC).

\strong{PK/PD target}:
The percent of time that the free concentration is above the minimum inhibitory concentration.
}
\examples{
\dontrun{
simulate_meropenem_mc_pta_ftime_above_mic(PATID = "Anonymous", 
    AGE = 65, HEIGHT = 175, 
    WEIGHT = 75, GENDER = "Male", 
    CREATININE = 1, MODEL = "Crandon et al. (2011) - ICU", 
    EDEMA = "No", PCTABOVEMIC = 40, 
    CRCLCAP = "No cap", 
    REGIMENS = list(list(
        DOSE = 500, INTERVAL = 6, 
        TINF = 0.5, set = "REGIMEN"), 
        list(DOSE = 1000, 
            INTERVAL = 8, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 2000, 
            INTERVAL = 8, 
            TINF = 0.5, 
            set = "REGIMEN"), 
        list(DOSE = 2000, 
            INTERVAL = 8, 
            TINF = 3, 
            set = "REGIMEN")))
}

}
\references{
\itemize{
  \item Jared L Crandon et al. (2011): Optimization of meropenem dosage in the critically ill population based on renal function. In. Intensive Care Medicine. https://pubmed.ncbi.nlm.nih.gov/21136037
  \item Li, C. et. al (2006): Population Pharmacokinetic Analysis and Dosing Regimen Optimization of Meropenem in Adult Patients. In. The Journal of Clinical Pharmacology. https://accp1.onlinelibrary.wiley.com/doi/10.1177/0091270006291035
  \item Doh, K. et al. (2010): Population pharmacokinetics of meropenem in burn patients. In. Journal of Antimicrobial Chemotherapy. https://academic.oup.com/jac/article/65/11/2428/762112
}
}
