% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_phenytoin_bayesian_ncmin.R
\name{simulate_phenytoin_bayesian_ncmin}
\alias{simulate_phenytoin_bayesian_ncmin}
\title{Phenytoin » Bayesian » Cmin}
\usage{
simulate_phenytoin_bayesian_ncmin(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  PHENYTOINSODIUM,
  MODEL,
  RENALFUNCTIONSTATUS,
  VALPROICACID,
  CMIN,
  ORALORIVHISTORYWITHALBUMIN,
  ORALORIVREGIMENS
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').}

\item{PHENYTOINSODIUM}{Phenytoin form. Form of phenytoin to be used during simulations. Must be provided as string ('Phenytoin sodium').}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Muñoz et al. (2019) - ICU').}

\item{RENALFUNCTIONSTATUS}{Renal function status. Renal function status of the patient. Must be provided as string ('Patients without renal impairment' or 'Dialysis or End-Stage Renal Disease').}

\item{VALPROICACID}{Valproic acid blood level (leave empty if none). Total valproic acid concentration, leave empty or set to zero if no concurrent administration. Must be provided as numeric (min. 0, max. 500 mg/L).}

\item{CMIN}{Goal corrected phenytoin trough level. Corrected phenytoin level is calculated using the May Equation given a measured valproic acid level, or with the help of Winter-Tozer Equation which involves the serum albumin level and renal function status. Must be provided as numeric (min. 5, max. 30 mg/L).}

\item{ORALORIVHISTORYWITHALBUMIN}{Historical Records.  Must be provided as list of 1-48 'HISTALBUMIN', 'HISTDOSE', 'HISTORALDOSE' or 'HISTCONCENTRATION' values.}

\item{ORALORIVREGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-20 'REGIMEN' or 'ORALREGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}
}
\description{
Phenytoin » Bayesian adaptive dosing » Normalized minimum concentration
}
\details{
\strong{Drug}:
Phenytoin

\strong{Method}:
Estimate the pharmacokinetic parameters of the patient from past concentrations with Bayesian inverse modeling, then use that information to predict the steady state concentrations for multiple dosing regimens and select the optimal one, with regard to the target pharmacodynamic index.

\strong{PK/PD target}:
Normalized minimum blood plasma concentration (mg/L).
}
\examples{
\dontrun{
structure(list("Anonymous", 
    45, 175, 75, "Male", 
    "Phenytoin sodium", 
    "Muñoz et al. (2019) - ICU", 
    "Dialysis or End-Stage Renal Disease", 
    NULL, 10, list(list(
        DATETIME = structure(1601866800, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
        ALBUMIN = 4.4, 
        set = "HISTALBUMIN"), 
        list(DATETIME = structure(1601870400, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            DOSE = 1500, 
            TINF = 0.5, 
            set = "HISTDOSE"), 
        list(DATETIME = structure(1601913600, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            CONCENTRATION = 13, 
            set = "HISTCONCENTRATION")), 
    list(list(set = "REGIMEN", 
        DOSE = 200, INTERVAL = 12, 
        TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 300, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 400, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 500, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 600, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 200, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 300, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 400, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 500, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 600, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "ORALREGIMEN", 
            DOSE = 200, 
            INTERVAL = 12), 
        list(set = "ORALREGIMEN", 
            DOSE = 300, 
            INTERVAL = 12), 
        list(set = "ORALREGIMEN", 
            DOSE = 400, 
            INTERVAL = 12), 
        list(set = "ORALREGIMEN", 
            DOSE = 500, 
            INTERVAL = 12), 
        list(set = "ORALREGIMEN", 
            DOSE = 600, 
            INTERVAL = 12), 
        list(set = "ORALREGIMEN", 
            DOSE = 200, 
            INTERVAL = 24), 
        list(set = "ORALREGIMEN", 
            DOSE = 300, 
            INTERVAL = 24), 
        list(set = "ORALREGIMEN", 
            DOSE = 400, 
            INTERVAL = 24), 
        list(set = "ORALREGIMEN", 
            DOSE = 500, 
            INTERVAL = 24), 
        list(set = "ORALREGIMEN", 
            DOSE = 600, 
            INTERVAL = 24))), .Names = c("PATID", 
"AGE", "HEIGHT", "WEIGHT", 
"GENDER", "PHENYTOINSODIUM", 
"MODEL", "RENALFUNCTIONSTATUS", 
NA, "CMIN", "ORALORIVHISTORYWITHALBUMIN", 
"ORALORIVREGIMENS"))
}

}
\references{
\itemize{
  \item Muñoz-Pichuante, Daniel et al. (2019): Dosage of phenytoin in neurocritical patients using Bayesian algorithms: a pilot study. In. Drug Metabolism and Personalized Therapy. https://pubmed.ncbi.nlm.nih.gov/31981450
  \item K. Soetaert, T. Petzoldt (2010): Inverse Modelling, Sensitivity and Monte Carlo Analysis in R Using Package FME. In. Journal of Statistical Software. https://www.jstatsoft.org/article/view/v033i03
}
}
