% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_phenytoin_mc_optloading_ncmin_24h.R
\name{simulate_phenytoin_mc_optloading_ncmin_24h}
\alias{simulate_phenytoin_mc_optloading_ncmin_24h}
\title{Phenytoin » Optimal loading dose » Cmin}
\usage{
simulate_phenytoin_mc_optloading_ncmin_24h(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  PHENYTOINFORM,
  MODEL,
  ALBUMIN,
  RENALFUNCTIONSTATUS,
  VALPROICACID,
  CPOSTDOSE24H
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Sex of the patient. Must be provided as string ('Male' or 'Female').}

\item{PHENYTOINFORM}{Phenytoin form. Form of phenytoin to be used during simulations. Must be provided as string ('Phenytoin sodium IV' or 'Fosphenytoin sodium IV').}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Muñoz et al. (2019) - ICU').}

\item{ALBUMIN}{Albumin. Serum albumin. Must be provided as numeric (min. 0.01, max. 15 g/dL).}

\item{RENALFUNCTIONSTATUS}{Renal function status. Renal function status of the patient. Must be provided as string ('Patients without renal impairment' or 'Dialysis or End-Stage Renal Disease').}

\item{VALPROICACID}{Valproic acid blood level (leave empty if none). Total valproic acid concentration, leave empty or set to zero if no concurrent administration. Must be provided as numeric (min. 0, max. 500 mg/L).}

\item{CPOSTDOSE24H}{Post-dose plasma concentration after 24 hrs target. The PK/PD target can be provided as post-dose blood plasma concentration after 24 hrs (Cmin). Must be provided as numeric (min. 1, max. 200 mg/L).}
}
\description{
Phenytoin » Optimal loading dose simulation » Post-dose concentration after 24 hrs
}
\section{Drug}{

Phenytoin
}

\section{Method}{

Calculate optimal loading dose and simulate concentrations for the first 24 hours.
}

\section{PK/PD target}{

Post-dose blood plasma concentration after 24 hrs (mg/L).
}

\examples{
\dontrun{
structure(list("Anonymous", 
    45, 175, 75, "Male", 
    "Fosphenytoin sodium IV", 
    "Muñoz et al. (2019) - ICU", 
    3, "Dialysis or End-Stage Renal Disease", 
    NULL, 10), .Names = c("PATID", 
"AGE", "HEIGHT", "WEIGHT", 
"GENDER", "PHENYTOINFORM", 
"MODEL", "ALBUMIN", "RENALFUNCTIONSTATUS", 
NA, "CPOSTDOSE24H"))
}
}
