% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/simulate_piperacillin_tazobactam_mc_pta_ftime_above_mic.R
\name{simulate_piperacillin_tazobactam_mc_pta_ftime_above_mic}
\alias{simulate_piperacillin_tazobactam_mc_pta_ftime_above_mic}
\title{Piperacillin & Tazobactam » PTA » % fT > MIC}
\usage{
simulate_piperacillin_tazobactam_mc_pta_ftime_above_mic(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  CREATININE,
  MODEL,
  PCTABOVEMIC,
  CRCLCAP,
  REGIMENS
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Sex of the patient. Must be provided as string ('Male' or 'Female').}

\item{CREATININE}{Creatinine. Serum creatinine. Must be provided as numeric (min. 0.01, max. 15 mg/dL).}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Felton et al. (2014) - ICU' or 'Patel et. al. (2010) - General ward').}

\item{PCTABOVEMIC}{Percent of time target that the drug concentration is above MIC. The PK/PD target can be provided as the percent of time that the drug concentration is above the minimum inhibitory concentration (% T > MIC). Must be provided as numeric (min. 5, max. 100 %).}

\item{CRCLCAP}{Capping Creatinine Clearance. Whether to use capping for creatinine clearance. Must be provided as string ('No cap', '120 ml/min', '130 ml/min', '140 ml/min' or '150 ml/min').}

\item{REGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-8 'REGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}
}
\description{
Piperacillin & Tazobactam » Probability of Target Attainment » Percent time of free concentration above MIC
}
\details{
\strong{Drug}:
Piperacillin & Tazobactam

\strong{Method}:
Monte Carlo simulation on the probabilities that a specific value of a pharmacodynamic index is achieved in case of different minimum inhibitory concentrations (MIC).

\strong{PK/PD target}:
The percent of time that the free concentration is above the minimum inhibitory concentration.
}
\examples{
\dontrun{
simulate_piperacillin_tazobactam_mc_pta_ftime_above_mic(PATID = "Anonymous", 
    AGE = 35, HEIGHT = 155, 
    WEIGHT = 55, GENDER = "Female", 
    CREATININE = 1.0638498, 
    MODEL = "Felton et al. (2014) - ICU", 
    PCTABOVEMIC = 50, 
    CRCLCAP = "No cap", 
    REGIMENS = list(list(
        DOSE = 3375, 
        INTERVAL = 6, 
        TINF = 0.5, set = "REGIMEN")))
}

}
\references{
\itemize{
  \item Felton, T. W. et al. (2014): Individualization of Piperacillin Dosing for Critically Ill Patients: Dosing Software To Optimize Antimicrobial Therapy. In. Antimicrobial Agents and Chemotherapy. https://aac.asm.org/content/58/7/4094
  \item Patel, N. et al. (2010): Identification of Optimal Renal Dosage Adjustments for Traditional and Extended-Infusion Piperacillin-Tazobactam Dosing Regimens in Hospitalized Patients. In. Antimicrobial Agents and Chemotherapy. https://aac.asm.org/content/54/1/460
}
}
