% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_voriconazole_bayesian_fauc_mic_ratio.R
\name{simulate_voriconazole_bayesian_fauc_mic_ratio}
\alias{simulate_voriconazole_bayesian_fauc_mic_ratio}
\title{Voriconazole » Bayesian » fAUC/MIC}
\usage{
simulate_voriconazole_bayesian_fauc_mic_ratio(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  MODEL,
  RIFAMPICIN,
  SEVEREHEPATICCHOLESTASIS,
  AUCPERMIC,
  ORALORIVBAREHISTORY,
  ORALORIVORKGREGIMENS
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Pascual et al. (2012) - patients with invasive fungal infections').}

\item{RIFAMPICIN}{Concurrent rifampicin. Presence of rifampicin coadministration. Must be provided as string ('No' or 'Yes').}

\item{SEVEREHEPATICCHOLESTASIS}{Severe hepatic cholestasis. Severe hepatic cholestasis is defined as any liver function tests > 20 times the upper limit of normal (e.g. alkaline phosphatase, AST, ALT or GGT). Must be provided as string ('No' or 'Yes').}

\item{AUCPERMIC}{AUC to MIC ratio target. The PK/PD target can be provided as 24 hour area under the concentration-time curve to minimum inhibitory concentration ratio (AUC/MIC). Must be provided as numeric (min. 10, max. 2000 ).}

\item{ORALORIVBAREHISTORY}{Historical Records.  Must be provided as list of 2-48 'HISTDOSE', 'HISTORALDOSE' or 'HISTCONCENTRATION' values.}

\item{ORALORIVORKGREGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-20 'KGREGIMEN', 'REGIMEN' or 'ORALREGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}
}
\description{
Voriconazole » Bayesian adaptive dosing » fAUC to MIC ratio
}
\details{
\strong{Drug}:
Voriconazole

\strong{Method}:
Estimate the pharmacokinetic parameters of the patient from past concentrations with Bayesian inverse modeling, then use that information to predict the steady state concentrations for multiple dosing regimens and select the optimal one, with regard to the target pharmacodynamic index.

\strong{PK/PD target}:
24 hour area under the free concentration-time curve to minimum inhibitory concentration ratio.
}
\examples{
\dontrun{
simulate_voriconazole_bayesian_fauc_mic_ratio(PATID = "Anonymous", 
    AGE = 35, HEIGHT = 155, 
    WEIGHT = 55, GENDER = "Female", 
    MODEL = "Pascual et al. (2012) - patients with invasive fungal infections", 
    RIFAMPICIN = "No", 
    SEVEREHEPATICCHOLESTASIS = "No", 
    AUCPERMIC = 25, ORALORIVBAREHISTORY = list(
        list(DATETIME = structure(1633579200, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            DOSE = 800, 
            set = "HISTORALDOSE"), 
        list(DATETIME = structure(1633622400, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            DOSE = 400, 
            TINF = 1.5, 
            set = "HISTDOSE"), 
        list(DATETIME = structure(1633593600, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            CONCENTRATION = 2.3, 
            set = "HISTCONCENTRATION"), 
        list(DATETIME = structure(1633665600, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            CONCENTRATION = 2, 
            set = "HISTCONCENTRATION")), 
    ORALORIVORKGREGIMENS = list(
        list(set = "REGIMEN", 
            DOSE = 100, 
            INTERVAL = 12, 
            TINF = 1.5), 
        list(set = "ORALREGIMEN", 
            DOSE = 100, 
            INTERVAL = 12), 
        list(set = "REGIMEN", 
            DOSE = 200, 
            INTERVAL = 12, 
            TINF = 1.5), 
        list(set = "ORALREGIMEN", 
            DOSE = 200, 
            INTERVAL = 12), 
        list(set = "REGIMEN", 
            DOSE = 300, 
            INTERVAL = 12, 
            TINF = 1.5), 
        list(set = "ORALREGIMEN", 
            DOSE = 300, 
            INTERVAL = 12)))
}

}
\references{
\itemize{
  \item Pascual, Andres et al. (2012): Challenging recommended oral and intravenous voriconazole doses for improved efficacy and safety: population pharmacokinetics-based analysis of adult patients with invasive fungal infections. In. Clinical Infectious Diseases. https://pubmed.ncbi.nlm.nih.gov/22610925
}
}
