% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_voriconazole_mc_cfr_fauc_mic_ratio.R
\name{simulate_voriconazole_mc_cfr_fauc_mic_ratio}
\alias{simulate_voriconazole_mc_cfr_fauc_mic_ratio}
\title{Voriconazole » CFR » fAUC/MIC}
\usage{
simulate_voriconazole_mc_cfr_fauc_mic_ratio(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  MODEL,
  RIFAMPICIN,
  SEVEREHEPATICCHOLESTASIS,
  AUCPERMIC,
  ORALORIVORKGREGIMENS,
  MICDISTR
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Pascual et al. (2012) - patients with invasive fungal infections').}

\item{RIFAMPICIN}{Concurrent rifampicin. Presence of rifampicin coadministration. Must be provided as string ('No' or 'Yes').}

\item{SEVEREHEPATICCHOLESTASIS}{Severe hepatic cholestasis. Severe hepatic cholestasis is defined as any liver function tests > 20 times the upper limit of normal (e.g. alkaline phosphatase, AST, ALT or GGT). Must be provided as string ('No' or 'Yes').}

\item{AUCPERMIC}{AUC to MIC ratio target. The PK/PD target can be provided as 24 hour area under the concentration-time curve to minimum inhibitory concentration ratio (AUC/MIC). Must be provided as numeric (min. 10, max. 2000 ).}

\item{ORALORIVORKGREGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-8 'KGREGIMEN', 'REGIMEN' or 'ORALREGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}

\item{MICDISTR}{MIC distribution. Non-cumulative MIC distribution of organism for the given antibiotic at the specified MICs provided as percentages with a sum of 100\%. MIC distribution of voriconazole when tested against Aspergillus fumigatus. [Borman et al. "MIC Distributions and Evaluation of Fungicidal Activity for Amphotericin B, Itraconazole, Voriconazole, Posaconazole and Caspofungin and 20 Species of Pathogenic Filamentous Fungi Determined Using the CLSI Broth Microdilution Method." Journal of fungi (Basel, Switzerland) vol. 3,2 27. 31 May. 2017, doi:10.3390/jof3020027]. Must be provided as set.}
}
\description{
Voriconazole » Cumulative Fraction of Response » fAUC to MIC ratio
}
\details{
\strong{Drug}:
Voriconazole

\strong{Method}:
Monte Carlo simulation on the proportion of the population achieving a certain pharmacodynamic index value, given the minimum inhibitory concentration (MIC) distribution of the target microorganism(s).

\strong{PK/PD target}:
24 hour area under the free concentration-time curve to minimum inhibitory concentration ratio.
}
\examples{
\dontrun{
simulate_voriconazole_mc_cfr_fauc_mic_ratio(PATID = "Anonymous", 
    AGE = 35, HEIGHT = 155, 
    WEIGHT = 55, GENDER = "Female", 
    MODEL = "Pascual et al. (2012) - patients with invasive fungal infections", 
    RIFAMPICIN = "No", 
    SEVEREHEPATICCHOLESTASIS = "No", 
    AUCPERMIC = 25, ORALORIVORKGREGIMENS = list(
        list(set = "REGIMEN", 
            DOSE = 200, 
            INTERVAL = 12, 
            TINF = 1.5), 
        list(set = "ORALREGIMEN", 
            DOSE = 200, 
            INTERVAL = 12)), 
    MICDISTR = list("MICDISTR-0.0625" = list(
        0.6), "MICDISTR-0.125" = list(
        13), "MICDISTR-0.25" = list(
        68.4), "MICDISTR-0.5" = list(
        12.5), "MICDISTR-1" = list(
        2.8), "MICDISTR-2" = list(
        1.9), "MICDISTR-4" = list(
        0.8)))
}

}
\references{
\itemize{
  \item Pascual, Andres et al. (2012): Challenging recommended oral and intravenous voriconazole doses for improved efficacy and safety: population pharmacokinetics-based analysis of adult patients with invasive fungal infections. In. Clinical Infectious Diseases. https://pubmed.ncbi.nlm.nih.gov/22610925
}
}
