% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_voriconazole_mc_emp_cmin_mic_ratio.R
\name{simulate_voriconazole_mc_emp_cmin_mic_ratio}
\alias{simulate_voriconazole_mc_emp_cmin_mic_ratio}
\title{Voriconazole » Empiric » Cmin/MIC}
\usage{
simulate_voriconazole_mc_emp_cmin_mic_ratio(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  MODEL,
  RIFAMPICIN,
  SEVEREHEPATICCHOLESTASIS,
  CMINPERMIC,
  LOADINGDOSE,
  ORALORIVORKGREGIMENS
)
}
\arguments{
\item{PATID}{Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.}

\item{AGE}{Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).}

\item{HEIGHT}{Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).}

\item{WEIGHT}{Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).}

\item{GENDER}{Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').}

\item{MODEL}{Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Pascual et al. (2012) - patients with invasive fungal infections').}

\item{RIFAMPICIN}{Concurrent rifampicin. Presence of rifampicin coadministration. Must be provided as string ('No' or 'Yes').}

\item{SEVEREHEPATICCHOLESTASIS}{Severe hepatic cholestasis. Severe hepatic cholestasis is defined as any liver function tests > 20 times the upper limit of normal (e.g. alkaline phosphatase, AST, ALT or GGT). Must be provided as string ('No' or 'Yes').}

\item{CMINPERMIC}{Minimum concentration to MIC ratio target. The PK/PD target can be provided as minimum concentration to minimum inhibitory concentration ratio (Cmin/MIC). Must be provided as numeric (min. 0.1, max. 50 ).}

\item{LOADINGDOSE}{Loading dose. Loading dose is desired or not. Must be provided as string ('No' or 'Yes').}

\item{ORALORIVORKGREGIMENS}{Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-20 'KGREGIMEN', 'REGIMEN' or 'ORALREGIMEN' values. Use the \code{regimen} helper function to define the REGIMEN values.}
}
\description{
Voriconazole » Empiric therapy » Total minimum concentration to MIC ratio
}
\details{
\strong{Drug}:
Voriconazole

\strong{Method}:
Simulate concentrations for multiple dosing regimens and select the optimal one, with regard to the target pharmacodynamic index.

\strong{PK/PD target}:
Minimum concentration (mg/L) to minimum inhibitory concentration ratio (mg/L).
}
\examples{
\dontrun{
simulate_voriconazole_mc_emp_cmin_mic_ratio(PATID = "Anonymous", 
    AGE = 35, HEIGHT = 155, 
    WEIGHT = 55, GENDER = "Female", 
    MODEL = "Pascual et al. (2012) - patients with invasive fungal infections", 
    RIFAMPICIN = "No", 
    SEVEREHEPATICCHOLESTASIS = "No", 
    CMINPERMIC = 3, LOADINGDOSE = "No", 
    ORALORIVORKGREGIMENS = list(
        list(set = "REGIMEN", 
            DOSE = 200, 
            INTERVAL = 12, 
            TINF = 1.5), 
        list(set = "ORALREGIMEN", 
            DOSE = 200, 
            INTERVAL = 12)))
}

}
\references{
\itemize{
  \item Pascual, Andres et al. (2012): Challenging recommended oral and intravenous voriconazole doses for improved efficacy and safety: population pharmacokinetics-based analysis of adult patients with invasive fungal infections. In. Clinical Infectious Diseases. https://pubmed.ncbi.nlm.nih.gov/22610925
}
}
