% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{simulate}
\alias{simulate}
\title{Hit Rx Studio simulation API}
\usage{
simulate(endpoint, ...)
}
\arguments{
\item{endpoint}{path under \url{https://api.rx.studio/v1/simulation/}}

\item{...}{named parameters passed}
}
\description{
Hit Rx Studio simulation API
}
\keyword{internal}
